requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pydantic==2.5.2
requests==2.31.0
python-dotenv==1.0.0
pygame==2.5.2
groq==0.3.0
apify-client==1.6.1
playht==0.1.3
